+++ packages/rules-engine/src/services/VariableService/variableService.types.js
@@ -20,7 +20,6 @@ type EventMain = {
     +programStageId?: string,
     +programStageName?: string,
     +orgUnitId?: string,
-    +orgUnitName?: string,
     +trackedEntityInstanceId?: string,
     +enrollmentId?: string,
     +enrollmentStatus?: string,