+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageDetail/StageDetail.component.js
@@ -104,6 +104,11 @@ const StageDetailPlain = (props: Props) => {
         onCreateNew(stageId);
     }, [onCreateNew, stageId]);
 
+    const handleShowMore = useCallback(() => {
+        const nextRowIndex = Math.min(events.length, displayedRowNumber + DEFAULT_NUMBER_OF_ROW);
+        setDisplayedRowNumber(nextRowIndex);
+    }, [events, displayedRowNumber, setDisplayedRowNumber]);
+
     function renderHeader() {
         const headerCells = headerColumns
             .map(column => (
@@ -182,16 +187,14 @@ const StageDetailPlain = (props: Props) => {
     }
 
     function renderFooter() {
-        const renderShowMoreButton = () => (events.length > DEFAULT_NUMBER_OF_ROW
+        const renderShowMoreButton = () => (dataSource && !loading
+            && events.length > DEFAULT_NUMBER_OF_ROW
             && displayedRowNumber < events.length ? <Button
                 small
                 secondary
                 dataTest="show-more-button"
                 className={classes.button}
-                onClick={() => {
-                    const nextRowIndex = Math.min(events.length, displayedRowNumber + DEFAULT_NUMBER_OF_ROW);
-                    setDisplayedRowNumber(nextRowIndex);
-                }}
+                onClick={handleShowMore}
             >
                 {i18n.t('Show {{ rest }} more', {
                     rest: Math.min(events.length - displayedRowNumber, DEFAULT_NUMBER_OF_ROW),