+++ src/core_modules/capture-core/flow/apiTypes.js
@@ -18,7 +18,6 @@ declare type ApiEnrollmentEvent = {|
     program: string,
     programStage: string,
     orgUnit: string,
-    orgUnitName: string,
     trackedEntityInstance: string,
     enrollment: string,
     enrollmentStatus: string,